Refactor Codes : add feature that create dialog when delete habit

+++ HabitBread/app/src/main/java/com/habitbread/main/ui/fragment/ModificationBottomSheet.kt
@@ -2,6 +2,7 @@ package com.habitbread.main.ui.fragment
 
 import android.app.Dialog
 import android.content.Context
+import android.content.DialogInterface
 import android.graphics.Color
 import android.os.Bundle
 import android.util.Log
@@ -12,6 +13,7 @@ import android.widget.EditText
 import android.widget.FrameLayout
 import android.widget.TextView
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import com.habitbread.main.R
@@ -21,6 +23,8 @@ import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.google.android.material.chip.Chip
 import com.google.android.material.switchmaterial.SwitchMaterial
+import com.habitbread.main.base.BaseApplication
+import com.habitbread.main.util.AccountUtils
 import kotlinx.android.synthetic.main.fragment_modification.*
 import ru.ifr0z.timepickercompact.TimePickerCompact
 
@@ -195,17 +199,38 @@ class ModificationBottomSheet : BottomSheetDialogFragment() {
         }
     }
 
+//    private fun onClickDelete() {
+//        textView_modify_delete.setOnClickListener {
+//            detailViewModel.deleteHabit(getHabitId)
+//            detailViewModel.deleteData.observe(this, Observer {
+//                if(it.message == "success") {
+//                    activity?.finish()
+//                    Toast.makeText(context, "습관빵이 삭제되었습니다.", Toast.LENGTH_SHORT).show()
+//                }else {
+//                    Toast.makeText(context, "죄송합니다. 오류로 인해 습관빵이 삭제되지 않았습니다.", Toast.LENGTH_SHORT).show()
+//                }
+//            })
+//        }
+//    }
+
     private fun onClickDelete() {
         textView_modify_delete.setOnClickListener {
-            detailViewModel.deleteHabit(getHabitId)
-            detailViewModel.deleteData.observe(this, Observer {
-                if(it.message == "success") {
-                    activity?.finish()
-                    Toast.makeText(context, "습관빵이 삭제되었습니다.", Toast.LENGTH_SHORT).show()
-                }else {
-                    Toast.makeText(context, "죄송합니다. 오류로 인해 습관빵이 삭제되지 않았습니다.", Toast.LENGTH_SHORT).show()
+            val dialog = AlertDialog.Builder(requireContext())
+                .setMessage("정말 삭제 하시겠습니까?")
+                .setPositiveButton("예") { dialogInterface: DialogInterface, i: Int ->
+                    detailViewModel.deleteHabit(getHabitId)
+                    detailViewModel.deleteData.observe(this, Observer {
+                        if(it.message == "success") {
+                            activity?.finish()
+                            Toast.makeText(context, "습관빵이 삭제되었습니다.", Toast.LENGTH_SHORT).show()
+                        }else {
+                            Toast.makeText(context, "죄송합니다. 오류로 인해 습관빵이 삭제되지 않았습니다.", Toast.LENGTH_SHORT).show()
+                        }
+                    })
+                }.setNegativeButton("아니요") { dialogInterface: DialogInterface, i: Int ->
+                    dialogInterface.dismiss()
                 }
-            })
+            dialog.create().show()
         }
     }
 }
